chore: refactoring - creating separate folder for context
