+++ packages/react-tinacms-editor/src/context/editorState.tsx
@@ -19,8 +19,8 @@ limitations under the License.
 import * as React from 'react'
 import { createContext, useContext, useEffect, useState } from 'react'
 import { EditorView } from 'prosemirror-view'
-import { buildEditor } from '../utils/buildEditor'
-import { ALL_PLUGINS } from '../../default-plugins'
+import { buildEditor } from './utils/buildEditor'
+import { ALL_PLUGINS } from '../default-plugins'
 
 const EditorStateContext = createContext<{
   editorView: { view: EditorView } | undefined
