+++ packages/react-tinacms-editor/src/context/utils/buildEditor.ts
